Refactor drag drop with new components
